+++ client/src/containers/Upload/Upload.scss
@@ -2,20 +2,6 @@
   margin-bottom: 5rem;
 }
 
-.three-column-layout {
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: row;
-  flex: 1;
-  align-items: flex-start;
-}
-
-.three-column-card {
-  margin: 0.5em;
-  width: 30%;
-  max-width: 30%;
-}
-
 .missing-info-table {
   width: 100%;
 }
