+++ client/src/contexts/RosterContext/RosterContext.tsx
@@ -6,8 +6,8 @@ import { stringify, parse } from 'query-string';
 import { useHistory } from 'react-router-dom';
 import UserContext from '../../contexts/UserContext/UserContext';
 import { User } from '../../shared/models';
-import { usePaginatedChildData } from '../../containers/Roster/hooks';
-import { getQueryMonthFormat } from '../../containers/Roster/rosterUtils';
+import { usePaginatedChildData } from '../../containers/RosterOld/hooks';
+import { getQueryMonthFormat } from '../../containers/RosterOld/rosterUtils';
 
 export const ALL_SITES = 'all-sites';
