+++ client/src/routes.ts
@@ -10,6 +10,7 @@ import SubmitSuccess from './containers/SubmitSuccess/SubmitSuccess';
 import AddRecord from './containers/CreateRecord/CreateRecord';
 import FundingSourceTimes from './containers/FundingSourceTimes/FundingSourceTimes';
 import Roster from './containers/Roster/Roster';
+// import Roster from './containers/RosterOld/Roster';
 import BatchEdit from './containers/BatchEdit/BatchEdit';
 import DevUtility from './containers/DevUtility/DevUtility';
 import PrivacyPolicy from './containers/PrivacyPolicy/PrivacyPolicy';
