+++ client/src/shared/payloads/AgeGroupCount.ts
@@ -1,0 +1,5 @@
+export type AgeGroupCount = {
+  'Infant/toddler': number, 
+  'Preschool': number,
+  'School aged': number,
+}
