+++ src/routes/children.ts
@@ -28,6 +28,7 @@ childrenRouter.get(
     }) as Moment;
     const skip = parseQueryString(req, 'skip', { post: parseInt }) as number;
     const take = parseQueryString(req, 'take', { post: parseInt }) as number;
+    const count = parseQueryString(req, 'count', { post: (countStr) => countStr === 'true' }) as boolean;
     const children = await controller.getActiveChildren(
       req.user,
       organizationIds,
@@ -37,6 +38,11 @@ childrenRouter.get(
         take,
       }
     );
+    if (count) {
+      const ageGroupCounts = await controller.getChildrenCountByAgeGroup(req.user, organizationIds);
+      res.send(ageGroupCounts);
+      return;
+    }
 
     res.send(children);
   })
